imp(CLI): uses clap to parse command line
This commit re-introduces `clap` as a dependency and removes `docopt`.
There are a few motivating factors:

- When `cargo-outdated` was originally created, it used `clap`
- When `cargo` became a dependency of `cargo-oudated` the CLI parser was
  changed to `docopt` because that's what `cargo` used.
- Since that time, `cargo` has changed it's CLI parser to `clap`
- [`docopt` has been effectively unmaintained for  years](https://github.com/clap-rs/clap/issues/2951#issuecomment-952882684), which is also noted in the README

This commit shaves ~31 seconds off the compile time on my CPU (Intel i9-9980HK). I
haven't dug into this, but I'm assuming because it's able to re-use the
`clap` compiled as for both `cargo-oudated` and `cargo` in the dep graph
(more on this below). It also seems to have removed some duplicate deps
with different versions.

Why use clap 2.x instead of clap 3.x or Structopt? Because that's what
`cargo` is using so we are able to re-use the compilation between. This
would still be the case with Structopt, however with clap v3 on the
horizon (fingers crossed) I'd prefer to just upgrade at that point
instead of worry about any differences. Plus any gains in compile time
would be reversed by using Structopt's Custom Derive (same with clap v3
eventually).