tests: adds cli ux tests for clap
This ensures that backwards compatibility was kept with previous
version, and also ensures it will be kept going forward.

One may ask why we're asserting that clap is working correctly? But
we're not, we're ensuring that the options we've told clap to use are
the working the way we expect. I.e. we're making sure we used the
correct options to get the desired outcome.